Windows: Refactor network modes
