+++ daemon/daemon_unix.go
@@ -17,6 +17,7 @@ import (
 	"github.com/docker/docker/daemon/graphdriver"
 	"github.com/docker/docker/pkg/archive"
 	"github.com/docker/docker/pkg/fileutils"
+	"github.com/docker/docker/pkg/parsers"
 	"github.com/docker/docker/pkg/parsers/kernel"
 	"github.com/docker/docker/pkg/system"
 	"github.com/docker/docker/runconfig"
@@ -459,4 +460,45 @@ func setupInitLayer(initLayer string) error {
 
 func (daemon *Daemon) NetworkApiRouter() func(w http.ResponseWriter, req *http.Request) {
 	return nwapi.NewHTTPHandler(daemon.netController)
+}
+
+func (daemon *Daemon) RegisterLinks(container *Container, hostConfig *runconfig.HostConfig) error {
+
+	if hostConfig == nil || hostConfig.Links == nil {
+		return nil
+	}
+
+	for _, l := range hostConfig.Links {
+		name, alias, err := parsers.ParseLink(l)
+		if err != nil {
+			return err
+		}
+		child, err := daemon.Get(name)
+		if err != nil {
+			//An error from daemon.Get() means this name could not be found
+			return fmt.Errorf("Could not get container for %s", name)
+		}
+		for child.hostConfig.NetworkMode.IsContainer() {
+			parts := strings.SplitN(string(child.hostConfig.NetworkMode), ":", 2)
+			child, err = daemon.Get(parts[1])
+			if err != nil {
+				return fmt.Errorf("Could not get container for %s", parts[1])
+			}
+		}
+		if child.hostConfig.NetworkMode.IsHost() {
+			return runconfig.ErrConflictHostNetworkAndLinks
+		}
+		if err := daemon.RegisterLink(container, child, alias); err != nil {
+			return err
+		}
+	}
+
+	// After we load all the links into the daemon
+	// set them to nil on the hostconfig
+	hostConfig.Links = nil
+	if err := container.WriteHostConfig(); err != nil {
+		return err
+	}
+
+	return nil
 }
