+++ runconfig/hostconfig_windows.go
@@ -1,0 +1,18 @@
+// +build !windows
+
+package runconfig
+
+func (n NetworkMode) IsDefault() bool {
+	return n == "default"
+}
+
+func DefaultDaemonNetworkMode() NetworkMode {
+	return NetworkMode("default")
+}
+
+func (n NetworkMode) NetworkName() string {
+	if n.IsDefault() {
+		return "default"
+	}
+	return ""
+}
