+++ runconfig/parse.go
@@ -24,6 +24,19 @@ var (
 	ErrConflictNetworkExposePorts       = fmt.Errorf("Conflicting options: --expose and the network mode (--expose)")
 )
 
+// validateNM is the set of fields passed to validateNetMode()
+type validateNM struct {
+	netMode      NetworkMode
+	flHostname   *string
+	flLinks      opts.ListOpts
+	flDns        opts.ListOpts
+	flExtraHosts opts.ListOpts
+	flMacAddress *string
+	flPublish    opts.ListOpts
+	flPublishAll *bool
+	flExpose     opts.ListOpts
+}
+
 func Parse(cmd *flag.FlagSet, args []string) (*Config, *HostConfig, *flag.FlagSet, error) {
 	var (
 		// FIXME: use utils.ListOpts for attach and volumes?
@@ -121,37 +134,22 @@ func Parse(cmd *flag.FlagSet, args []string) (*Config, *HostConfig, *flag.FlagSe
 		return nil, nil, cmd, fmt.Errorf("--net: invalid net mode: %v", err)
 	}
 
-	if (netMode.IsHost() || netMode.IsContainer()) && *flHostname != "" {
-		return nil, nil, cmd, ErrConflictNetworkHostname
+	vals := validateNM{
+		netMode:      netMode,
+		flHostname:   flHostname,
+		flLinks:      flLinks,
+		flDns:        flDns,
+		flExtraHosts: flExtraHosts,
+		flMacAddress: flMacAddress,
+		flPublish:    flPublish,
+		flPublishAll: flPublishAll,
+		flExpose:     flExpose,
 	}
 
-	if netMode.IsHost() && flLinks.Len() > 0 {
-		return nil, nil, cmd, ErrConflictHostNetworkAndLinks
+	if err := validateNetMode(&vals); err != nil {
+		return nil, nil, cmd, err
 	}
 
-	if netMode.IsContainer() && flLinks.Len() > 0 {
-		return nil, nil, cmd, ErrConflictContainerNetworkAndLinks
-	}
-
-	if (netMode.IsHost() || netMode.IsContainer()) && flDns.Len() > 0 {
-		return nil, nil, cmd, ErrConflictNetworkAndDns
-	}
-
-	if (netMode.IsContainer() || netMode.IsHost()) && flExtraHosts.Len() > 0 {
-		return nil, nil, cmd, ErrConflictNetworkHosts
-	}
-
-	if (netMode.IsContainer() || netMode.IsHost()) && *flMacAddress != "" {
-		return nil, nil, cmd, ErrConflictContainerNetworkAndMac
-	}
-
-	if netMode.IsContainer() && (flPublish.Len() > 0 || *flPublishAll == true) {
-		return nil, nil, cmd, ErrConflictNetworkPublishPorts
-	}
-
-	if netMode.IsContainer() && flExpose.Len() > 0 {
-		return nil, nil, cmd, ErrConflictNetworkExposePorts
-	}
 	// Validate the input mac address
 	if *flMacAddress != "" {
 		if _, err := opts.ValidateMACAddress(*flMacAddress); err != nil {
@@ -461,20 +459,6 @@ func parseKeyValueOpts(opts opts.ListOpts) ([]KeyValuePair, error) {
 		out[i] = KeyValuePair{Key: k, Value: v}
 	}
 	return out, nil
-}
-
-func parseNetMode(netMode string) (NetworkMode, error) {
-	parts := strings.Split(netMode, ":")
-	switch mode := parts[0]; mode {
-	case "default", "bridge", "none", "host":
-	case "container":
-		if len(parts) < 2 || parts[1] == "" {
-			return "", fmt.Errorf("invalid container format container:<name|id>")
-		}
-	default:
-		return "", fmt.Errorf("invalid --net: %s", netMode)
-	}
-	return NetworkMode(netMode), nil
 }
 
 func ParseDevice(device string) (DeviceMapping, error) {
