+++ runconfig/parse_windows.go
@@ -1,0 +1,20 @@
+package runconfig
+
+import (
+	"fmt"
+	"strings"
+)
+
+func parseNetMode(netMode string) (NetworkMode, error) {
+	parts := strings.Split(netMode, ":")
+	switch mode := parts[0]; mode {
+	case "default":
+	default:
+		return "", fmt.Errorf("invalid --net: %s", netMode)
+	}
+	return NetworkMode(netMode), nil
+}
+
+func validateNetMode(vals *validateNM) error {
+	return nil
+}
